major changes
added verifytoken end point for veryfying the localstorage item, and not just assign it without checking if token is vlaid.

this included:
creating an endpoint in the server, creating a function in the useAuth custom hook to call it, and calling that function on useAuth component mount.

in addition, added the setAuthToken helper fucntion so that i dont have to specidy the authorization everytime in the useJobs custom hook.

added checks in dashboard and navbar if user.token is present.